(re-)Add support for custom file deletion command (was cron_shred/cron_shred_command)
git-svn-id: https://subversion.assembla.com/svn/file_sender/filesender/branches/filesender-2.0@3257 8555b757-d854-4b86-925a-82fd84c90ff4

+++ classes/storage/StorageFilesystem.class.php
@@ -343,8 +343,19 @@ class StorageFilesystem {
         
         if(!file_exists($file_path)) return;
         
-        if(!unlink($file_path))
-            throw new StorageFilesystemCannotDeleteException($file_path);
+        $rm_command = Config::get('rm_command');
+        
+        if($rm_command) {
+            $cmd = str_replace('{path}', escapeshellarg($file_path), $rm_command);
+            exec($cmd, $out, $ret);
+            
+            if($ret)
+                throw new StorageFilesystemCannotDeleteException($file_path);
+            
+        } else {
+            if(!unlink($file_path))
+                throw new StorageFilesystemCannotDeleteException($file_path);
+        }
     }
     
     /**
